changes in migration

+++ database/migrations/2024_12_06_140807_add_adjustor_meeting_id_into_adjustor_square_photos_table.php
@@ -1,8 +1,9 @@
 <?php
 
-use Illuminate\Database\Migrations\Migration;
-use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Schema;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
 
 return new class extends Migration
 {
@@ -14,13 +15,17 @@ return new class extends Migration
     public function up()
     {
         Schema::table('adjustor_square_photos', function (Blueprint $table) {
-            $table->unsignedBigInteger('adjustor_meeting_id')->nullable()->after('id');
+            // $table->unsignedBigInteger('adjustor_meeting_id')->nullable()->after('id');
 
-            // Add the foreign key constraint
-            $table->foreign('adjustor_meeting_id')
-                ->references('id')
-                ->on('adjustor_meetings')
-                ->onDelete('cascade');        });
+            // // Add the foreign key constraint
+            // $table->foreign('adjustor_meeting_id')
+            //     ->references('id')
+            //     ->on('adjustor_meetings')
+            //     ->onDelete('cascade');        
+            if (!DB::getSchemaBuilder()->hasColumn('adjustor_square_photos', 'adjustor_meeting_id')) {
+                $table->unsignedBigInteger('adjustor_meeting_id')->nullable()->after('id');
+            }
+        });
     }
 
     /**
@@ -31,9 +36,11 @@ return new class extends Migration
     public function down()
     {
         Schema::table('adjustor_square_photos', function (Blueprint $table) {
-            $table->dropForeign(['adjustor_meeting_id']);
-            $table->dropColumn('adjustor_meeting_id');
-
+            // $table->dropForeign(['adjustor_meeting_id']);
+            // $table->dropColumn('adjustor_meeting_id');
+            if (DB::getSchemaBuilder()->hasColumn('adjustor_square_photos', 'adjustor_meeting_id')) {
+                $table->dropColumn('adjustor_meeting_id');
+            }
         });
     }
 };
